refactor(oauth): move /verify route
